# Conflicts:
#	resources/views/admin/layouts/master1.blade.php
#	routes/web.php

+++ app/Http/Controllers/BookController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\BookCreateRequest;
 use App\Models\Book;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Session;
@@ -14,27 +16,26 @@ class BookController extends Controller
     {
         $books = Book::all();
 
-        return view('books.list', compact('books'));
+        return view('backends.admin.books.list', compact('books'));
     }
 
     public function create()
     {
-        return view('books.create');
+        return view('backends.admin.books.create');
     }
 
-    public function store(Request $request, Book $book): \Illuminate\Http\RedirectResponse
+    public function cre(BookCreateRequest $request, Book $book): \Illuminate\Http\RedirectResponse
     {
-
-        $image = $request->file('image');
-        $path = $image->store('images', 'public');
-        $book->image = $path;
-
+        if ($request->hasFile('image')) {
+            $image = $request->file('image');
+            $path = $image->store('images', 'public');
+            $book->image = $path;
+        }
         $book->name = $request->name;
-        $book->image = $request->image;
-        $book->author = $request->author;
-        $book->category = $request->category;
-        $book->price = $request->price;
-        $book->description = $request->description;
+        $book->author_id = $request->author;
+        $book->category_id = $request->category;
+        $book->year_publication = $request->year_publication;
+        $book->content = $request->contents;
         $book->save();
 
         Session::flash('success', 'success');
@@ -45,33 +46,34 @@ class BookController extends Controller
     public function detail($id)
     {
         $book = DB::table('books')->where('id', $id)->get();
-        return view('books.detail', compact('book'));
+        return view('backends.admin.books.detail', compact('book'));
     }
 
     public function edit($id)
     {
         $book = Book::findOrFail($id);
-        return view('books.edit', compact('book'));
+        return view('backends.admin.books.edit', compact('book'));
     }
 
-    public function update(Request $request, $id)
+    public function update(Request $request, $id): RedirectResponse
     {
         $book = Book::findOrFail($id);
-        $currentImg = $book->image;
-        if ($currentImg) {
-            Storage::delete('/public/' . $currentImg);
-        }
-        $image = $request->file('image');
-        $path = $image->store('images', 'public');
-        $book->image = $path;
 
+        if ($request->hasFile('image')) {
+            $currentImg = $book->image;
+            if ($currentImg) {
+                Storage::delete('/public/' . $currentImg);
+            }
+            $image = $request->file('image');
+            $path = $image->store('images', 'public');
+            $book->image = $path;
+        }
 
         $book->name = $request->name;
-        $book->image=$request->image;
-        $book->author = $request->author;
-        $book->category = $request->category;
-        $book->price = $request->price;
-        $book->description = $request->description;
+        $book->author_id = $request->author;
+        $book->category_id = $request->category;
+        $book->year_publication = $request->year_publication;
+        $book->content = $request->contents;
         $book->save();
 
         Session::flash('success', 'success');
